add view model for statistics

+++ CarManager/Screens/Additional Views/PieChartView.swift
@@ -9,7 +9,7 @@ import Foundation
 import UIKit
 
 class PieChartView: UIView {
-    private var data: [(String, CGFloat)] = []
+    private var data: [(String, Int)] = []
     private var colors: [UIColor] = [.clear]
     private var strokeWidth: CGFloat = 1
     private var borderColor: UIColor = .black
@@ -19,12 +19,12 @@ class PieChartView: UIView {
         fatalError("\(Self.self) \(#function) has not been implemented")
     }
     
-    func setData(data: [(String, Double)], colors: [UIColor]) {
+    func setData(data: [(String, Int)], colors: [UIColor]) {
         let sum = data.reduce(into: 0) { $0 += $1.1 }
         if sum == 0 {
-            self.data = data.map { ($0.0, CGFloat($0.1)) }
+            self.data = data.map { ($0.0, $0.1) }
         } else {
-            self.data = data.map { ($0.0, CGFloat($0.1 / sum)) }.sorted { $0.1 > $1.1 }
+            self.data = data.map { ($0.0, $0.1 / sum) }.sorted { $0.1 > $1.1 }
         }
         self.colors = colors.map { $0.withAlphaComponent(0.8) }
         setNeedsDisplay()
@@ -62,7 +62,7 @@ class PieChartView: UIView {
             return
         }
         for (index, (_, value)) in data.enumerated() {
-            let valueAngle = value * (2 * CGFloat.pi)
+            let valueAngle = CGFloat(value) * 2.0 * CGFloat.pi
             
             drawContext.saveGState()
             drawContext.translateBy(x: pieCenter.x, y: pieCenter.y)
@@ -71,7 +71,7 @@ class PieChartView: UIView {
             let path = UIBezierPath()
             path.move(to: CGPoint.zero)
             path.addLine(to: CGPoint(x: mainRadius, y: 0))
-            path.addArc(withCenter: CGPoint.zero, radius: mainRadius, startAngle: 0, endAngle: valueAngle, clockwise: true)
+            path.addArc(withCenter: CGPoint.zero, radius: mainRadius, startAngle: 0, endAngle: CGFloat(valueAngle), clockwise: true)
             path.close()
             
             drawContext.addPath(path.cgPath)
